refactor(widget): fold single-use files into their only consumers
Two behavior-preserving single-use file folds:

- constants/theme.ts (DARK_THEME_COLORS, 1 importer) -> inlined at the
  top of its sole consumer WidgetSettingsLoader.tsx; removes the now-empty
  src/constants/ directory.
- base/tokens.ts (spacingScale + SpacingToken, 1 importer) -> moved into
  its sole consumer base/layoutProps.ts (SpacingToken kept exported as a
  public LayoutProps prop type; spacingScale is now file-private).

No synced surface touched (src/sdk/* untouched); no npm publish.
